+++ src/snowflake/snowpark/_internal/data_source/utils.py
@@ -549,3 +549,21 @@ def process_parquet_queue_with_threads(
     )
 
     return fetch_to_local_end_time, upload_to_sf_start_time, upload_to_sf_end_time
+
+
+def track_data_source_statement_params(
+    dataframe, statement_params: Optional[Dict] = None
+) -> Optional[Dict]:
+    """
+    Helper method to initialize and update data source tracking statement_params based on dataframe attributes.
+    """
+    statement_params = statement_params or {}
+    if (
+        dataframe._plan
+        and dataframe._plan.api_calls
+        and dataframe._plan.api_calls[0].get("name") == DATA_SOURCE_DBAPI_SIGNATURE
+    ):
+        # Track data source ingestion
+        statement_params[STATEMENT_PARAMS_DATA_SOURCE] = "1"
+
+    return statement_params if statement_params else None